+++ akka-cluster-sharding/src/main/scala/akka/cluster/sharding/internal/LeastShardAllocationStrategy.scala
@@ -43,15 +43,6 @@ import akka.cluster.sharding.internal.AbstractLeastShardAllocationStrategy.Shard
     extends AbstractLeastShardAllocationStrategy {
   import LeastShardAllocationStrategy.emptyRebalanceResult
 
-  override def allocateShard(
-      requester: ActorRef,
-      shardId: ShardId,
-      currentShardAllocations: Map[ActorRef, immutable.IndexedSeq[ShardId]]): Future[ActorRef] = {
-    val regionEntries = regionEntriesFor(currentShardAllocations)
-    val (region, _) = mostSuitableRegion(regionEntries)
-    Future.successful(region)
-  }
-
   override def rebalance(
       currentShardAllocations: Map[ActorRef, immutable.IndexedSeq[ShardId]],
       rebalanceInProgress: Set[ShardId]): Future[Set[ShardId]] = {
